requirements.txt
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
Pillow==8.4.0
PyJWT==2.1.0
PySocks==1.7.1
Werkzeug==2.0.2
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.3
cryptography==3.4.8
cycler==0.11.0
flatbuffers==2.0
fonttools==4.28.5
frozenlist==1.2.0
gast==0.3.3
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.42.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.8.2
keras==2.7.0
kiwisolver==1.3.2
libclang==12.0.0
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pip==21.2.4
protobuf==3.17.2
pyOpenSSL==21.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.0.6
pyreadline==2.1
python-chess==0.31.3
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
setuptools==58.0.4
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.13.3
yarl==1.6.3
zipp==3.6.0